client/src/arduino_led/pyproject.toml
[project]
name = "arduino-led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
]
